Move BsPixApplication to app package.
