+++ app/src/debug/java/com/bsstokes/bspix/di/ComponentFactory.java
@@ -2,7 +2,7 @@ package com.bsstokes.bspix.di;
 
 import android.support.annotation.NonNull;
 
-import com.bsstokes.bspix.BsPixApplication;
+import com.bsstokes.bspix.app.BsPixApplication;
 
 public class ComponentFactory {
     public static AppComponent create(@NonNull BsPixApplication application) {
