+++ app/src/main/java/com/bsstokes/bspix/app/BsPixApplication.java
@@ -1,4 +1,4 @@
-package com.bsstokes.bspix;
+package com.bsstokes.bspix.app;
 
 import android.app.Activity;
 import android.app.Application;
